+++ changelog.rst
@@ -1,10 +1,14 @@
+2.0.0
+=======
+
+- Add react-paginate, react hook useState
+
 1.0.0
 =======
 
 - First version
-- Fetch API, get pokemon list 
-- toggleViewMode (list or grid)
-- remember toggleViewMode
+- Fetch API, get pokemon list, react hook useState
+- toggleViewMode (list or grid), react hook useState
 - grid (4 col pc, 3 col tablet, 2 smartphone) 
 - responsive CSS